Fix bug where new games carry over old extended questions
This was being caused by the way the questions and arguments arrays
were being mutated. This isn't a great way of handling this logic,
and my current plan is to do a big refactor of this logic to
try to simplify things. Storing them in state and coupling the
logic to set questions and scores together makes more sense here.