Only dynamically import ScoreBoard
Dynamically importing the whole page file and excluding it from SSR was overkill. The only issue is with the ScoreBoard
where the local storage data is stored. We can be much more granular and exclude that component from SSR in order to
still reap a lot of benefit from Next's prebuild.